@@ -13,15 +13,17 @@
         }
     ],
     "require": {
-		"php": "^7.4",
-        "symfony/process": "^5.1"
+		"php": "^7.4|^8.0",
+        "symfony/process": "^5.2.0",
+        "ext-json": "*"
     },
 	"require-dev": {
-        "dms/phpunit-arraysubset-asserts": "^0.2.0",
-        "friendsofphp/php-cs-fixer": "^2.16",
-        "pestphp/pest": "^0.1.5",
-        "phpunit/phpunit": "^9.0",
-        "vimeo/psalm": "^3.11"
+        "dms/phpunit-arraysubset-asserts": "^0.2.1",
+        "friendsofphp/php-cs-fixer": "^2.17.1",
+        "pestphp/pest": "^0.3.15",
+        "phpunit/phpunit": "^9.5",
+        "phpstan/phpstan": "0.12.59",
+        "ext-json": "*"
     },
 	"autoload": {
         "psr-4": {
@@ -34,11 +36,14 @@
         }
     },
     "scripts": {
-        "psalm": "vendor/bin/psalm",
-        "test": "vendor/bin/pest",
-        "format": "vendor/bin/php-cs-fixer fix --allow-risky=yes",
-        "post-install-cmd": [
-            "npm install"
+        "lint": "php-cs-fixer fix -v",
+        "test:lint": "php-cs-fixer fix -v --dry-run",
+        "test:types": "phpstan analyse --ansi --memory-limit=0",
+        "test:unit": "pest --colors=always",
+        "test": [
+            "@test:lint",
+            "@test:types",
+            "@test:unit"
         ]
     },
     "config": {
