@@ -4,16 +4,16 @@ namespace Octoper\Lighthouse\Exceptions;
 
 class AuditFailedException extends \Exception
 {
-    protected $output;
+    protected ?string $output;
 
-    public function __construct($url, $output = null)
+    public function __construct(string $url, string $output = null)
     {
         parent::__construct("Audit of '{$url}' failed");
 
         $this->output = $output;
     }
 
-    public function getOutput()
+    public function getOutput(): ?string
     {
         return $this->output;
     }
